pyproject.toml
[project]
name = "flowforge"
version = "0.1.0"
description = "FlowForge - Multi-Agent AI Data Pipeline Platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "confluent-kafka>=2.3.0",
    "psycopg2-binary>=2.9.9",
    "pymongo>=4.6.0",
    "redis>=5.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "uvicorn>=0.27.0",
    "fastapi>=0.109.0",
    "websockets>=12.0",
    "openai>=1.12.0",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
flowforge = "flowforge.cli:app"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["flowforge*"]

[tool.ruff]
line-length = 100
target-version = "py310"
